First Project of CS 300
In this project, a random maze generator and a path finder which can find a path in the maze which is created virtually. At the beginning, it asks the size of the maze and number of the mazes that will be generated. It generates exact amount of mazes and their sizes are same with what user gave. After that user is asked chose one of the mazes which are in txt format, user is also asked to chose an antry point to the maze as X,Y coordinates and exit points as X.Y coordinates. Then, path finder algorithm runs and finds a path and writes the found path into a txt folder as X,Y coordinate step-by-step. 

By using mazeDrawer, user can illustrate the any maze file which is written as txt. İmport txt file into mazeDrawer folder and run mazeDrawer.exe or mazeDrawer.linux or mazeDrawer.mac file based on their OS. After running the file, following the instructions will be enough and a mazeDrawn.txt file will be created.